# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GraphEdge": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "source": {
            "format": "uuid",
            "title": "Source",
            "type": "string"
          },
          "target": {
            "format": "uuid",
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "target"
        ],
        "title": "GraphEdge",
        "type": "object"
      },
      "GraphNode": {
        "additionalProperties": false,
        "description": "One agent on the canvas. `id` is generated by the frontend (crypto.randomUUID()).",
        "properties": {
          "agent_type": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Agent Type",
            "type": "string"
          },
          "configuration": {
            "additionalProperties": true,
            "title": "Configuration",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "requires_approval": {
            "default": false,
            "title": "Requires Approval",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "agent_type"
        ],
        "title": "GraphNode",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Position": {
        "properties": {
          "x": {
            "default": 0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "default": 0,
            "title": "Y",
            "type": "number"
          }
        },
        "title": "Position",
        "type": "object"
      },
      "RunCreated": {
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunCreated",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "awaiting_approval",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunSummary": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "edge_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edge Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "severity": {
            "default": "error",
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "ValidationResult": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "issues"
        ],
        "title": "ValidationResult",
        "type": "object"
      },
      "WorkflowCreate": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/WorkflowGraph"
          },
          "name": {
            "default": "Untitled workflow",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "WorkflowCreate",
        "type": "object"
      },
      "WorkflowGraph": {
        "additionalProperties": false,
        "description": "What the canvas saves into `workflows.graph_definition`. Never holds secrets.",
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "WorkflowGraph",
        "type": "object"
      },
      "WorkflowOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "graph": {
            "$ref": "#/components/schemas/WorkflowGraph"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "graph",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowOut",
        "type": "object"
      },
      "WorkflowSummary": {
        "properties": {
          "agent_types": {
            "items": {
              "type": "string"
            },
            "title": "Agent Types",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "agent_types",
          "updated_at"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      },
      "WorkflowUpdate": {
        "properties": {
          "graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowGraph"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "WorkflowUpdate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Visual AI-agent workflow platform \u2014 C2 api. Contract agreed in GP-plan \u00a75.3.",
    "title": "GP Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workflows": {
      "get": {
        "operationId": "list_workflows_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowSummary"
                  },
                  "title": "Response List Workflows Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{workflow_id}": {
      "delete": {
        "operationId": "delete_workflow_workflows__workflow_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Workflow",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "get_workflow_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "operationId": "update_workflow_workflows__workflow_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{workflow_id}/run": {
      "post": {
        "description": "Enqueue a run and return immediately with 202 \u2014 the run has not finished (NFR-01).",
        "operationId": "run_workflow_workflows__workflow_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "The workflow has validation errors"
          },
          "503": {
            "description": "Job queue unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{workflow_id}/runs": {
      "get": {
        "operationId": "list_runs_workflows__workflow_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Workflows  Workflow Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{workflow_id}/validate": {
      "post": {
        "operationId": "validate_workflow_workflows__workflow_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Workflow",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}